hackathons done

+++ lib/features/hackathons/models/hackathon_models.dart
@@ -6,15 +6,15 @@ part 'hackathon_models.g.dart';
 @freezed
 class HackathonModels with _$HackathonModels {
   factory HackathonModels({
-    String? id,
+    String? id, 
     @JsonKey(name: 'user_id') String? userId,
     String? title,
     String? theme,
     String? description,
     String? date,
     String? location,
-    @JsonKey(name: 'entry_fee') int? entryFee,
-    @JsonKey(name: 'prize_pool') int? prizePool,
+    @JsonKey(name: 'entry_fee') double? entryFee,
+    @JsonKey(name: 'prize_pool') double? prizePool,
     String? media,
     @JsonKey(name: 'registration_deadline') String? registrationDeadline,
     @JsonKey(name: 'organizer_name') String? organizerName,
